Trying to reset admin_password to the original one
For a newly launch we need to pass initial_password=XXX and ansible_password is
picked up from that value. At the end of this role we change to a permanent
admin password from vault. This atempts to set ansible_password to the one from
vault so the next role ansible will have correct password.