Make certificate import option exportable to yes
ansible saying if not supplied it is defaulted to yes.

However I do not know why the xvt.technology certificate imported can not be
attached to the IIS with error described in this article
https://knowledge.geotrust.com/support/knowledge-base/index?page=content&id=SO22115&actp=search&viewlocale=en_US&searchid=1435162914930.

I did a manual fix using that article so obviously it is the reason.

Now trying to explicitly set to yes if no supplied to see if issue is fixed.