Fixing the import certificate scripts
In theory the current should work. win_shell should take line by line and
execute it. However doing like that generate errors as of now which complains
about the password is not matched.

Just changing directly execute like that into deploying as a small import shell
script and run the import fix the issue.

Also this will save a copy of the script in the remote server for reference
thus it is a good thing to do as well.