allow-set-cert-location
This change allow to set the cert_store_location and allow us to import non pfx
certificates as well.

For non pfx we do nto support password cert yet as we are not sure if it is
well supported by powershell to import both private key or public key. For now
tested with .cer and only public key.

So if you want to import a cert with private and public key for web binding for
example, convert them to pfx for sure.

Use .cer if you only want to import a public key - like a issuer CA certificates.